requirements.txt
flask==2.0.2
jinja2==3.0.2
gunicorn==20.1.0
python-dotenv==0.19.2
Flask-Minify==0.37
requests==2.31.0